+++ native_library/ios/AdMobANE/AdMobANE/Extensions/FreNSError.swift
@@ -13,15 +13,11 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package com.tuarua.admobane
 
-data class AdMobEvent(val position: Int, val errorCode: Int = 0)
-data class AdMobEventWithReward(val position: Int, val errorCode: Int = 0,
-                                val amount: Int = 0, val type: String? = null)
+import Foundation
 
-data class ConsentEvent(val callbackId: String, val data: Map<String, Any>? = null, val error: Map<String, Any>? = null) {
-    companion object {
-        const val ON_CONSENT_INFO_UPDATE = "AdMob.onConsentInfoUpdate"
-        const val ON_CONSENT_FORM_DISMISSED = "AdMob.onConsentFormDismissed"
+public extension NSError {
+    func toDictionary() -> [String: Any] {
+        return ["text": self.localizedDescription, "id": self.code]
     }
 }